requirements.txt
pandas==2.0.3
numpy==1.24.3
tweepy==4.14.0
python-dotenv==1.0.0
vaderSentiment==3.3.2
transformers==4.30.2
torch==2.0.1
yfinance==0.2.28
plotly==5.14.0
streamlit==1.24.1
scikit-learn==1.3.0
datasets==2.13.1
tqdm==4.65.0
scipy==1.11.1
nbformat==5.9.2
ipykernel==6.25.1
